+++ src/shared/index.ts
@@ -1,4 +1,3 @@
 export * from "./components";
 export * from "./contexts";
-export * from "./services";
 export * from "./utils";
